refactor: split big media api to small modules
